% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{calc_Diversity}
\alias{calc_Diversity}
\alias{calc_Diversity.session}
\alias{calc_Diversity.Data}
\alias{Shannon_Entropy}
\alias{Percent_Complexity}
\alias{Nucleotide_Diversity}
\alias{Percent_Diversity}
\title{calc_Diversity generic}
\usage{
calc_Diversity(x)

\method{calc_Diversity}{session}(x)

\method{calc_Diversity}{Data}(x)

Shannon_Entropy(x)

Percent_Complexity(x)

Nucleotide_Diversity(x)

Percent_Diversity(x)
}
\arguments{
\item{x}{Data object or session object containing Data objects to calculate diversity in.}
}
\value{
Returns dataframe of calculated sequence-based diversity measures.
}
\description{
Calculates four sequence-based diversity measures in NGS Data objects.
}
