% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{close_session}
\alias{close_session}
\title{Remove session object from global environment.}
\usage{
close_session(session)
}
\arguments{
\item{session}{\code{session} object or character string name of \code{session} object to be removed. Default is "highlineR.session".}
}
\value{
Removes specified \code{session} object from global R environment.
}
\description{
\code{close_session} removes a session object from the global R environment.
}
