% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{compress}
\alias{compress}
\alias{compress.session}
\alias{compress.Data}
\alias{compress.default}
\alias{read_sample}
\alias{resample}
\alias{resample.session}
\alias{resample.Data}
\title{Compress generic}
\usage{
compress(x, N, M, unique, force)

\method{compress}{session}(x, N = 500, M = 10, unique = F,
  force = F)

\method{compress}{Data}(x, N = 500, M = 10, unique = F, force = F)

\method{compress}{default}(x, N, M, unique, force)

read_sample(x, N, M)

resample(x, N, M)

\method{resample}{session}(x, N = 500, M = 10)

\method{resample}{Data}(x, N = 500, M = 10)
}
\arguments{
\item{x}{Data object or session object containing Data objects to compress.}

\item{N}{Integer expected number of reads.}

\item{M}{Integer number of random samples to select from dataset for read averaging.}

\item{unique}{A logical value specifying whether the NGS data file contains unique sequences. If \code{True}, then it is assumed that read counts are present in sequence labels. If \code{False}, then read counts are inferred from the number of distinct instances of sequences in the file.}

\item{force}{A logical value specifying if previously imported file should be overwritten and re-imported.}
}
\value{
Compresses Data objects by populating the object's \code{compressed} environment with (sequence: abundance) key:value pairs.

\code{read_sample} samples \code{N} number of sequences from the compressed evironment M times and stores the average result in the Data object's \code{sample} environment.

\code{resample} resamples Data objects by overwriting the object's \code{sample} environment with (sequence: abundance) key:value pairs without re-compressing.
}
\description{
Compresses the NGS file specified by the path variable of \code{Data} objects.

\code{read_sample} samples \code{N} number of sequences from the compressed environment M times.

\code{resample} resamples Data objects without re-compressing.
}
