% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{convert_quality}
\alias{convert_quality}
\title{Convert encoded quality scores}
\usage{
convert_quality(line, encoding = "sanger", ...)
}
\arguments{
\item{line}{Character string of encoded quality scores.}

\item{encoding}{Character string of quality score encoding. Options: "Sanger" (default), "Solexa", "Illumina1.3", "Illumina1.5". "Illumina1.8".}
}
\value{
Numeric vector of converted quality scores.
}
\description{
\code{convert_quality} converts encoded quality scores.
}
\seealso{
\link{parse}
}
