% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_file}
\alias{import_file}
\title{Import NGS file.}
\usage{
import_file(path, datatype, seqtype, session, force = FALSE)
}
\arguments{
\item{path}{A character string representing the absolute path to sequence file.}

\item{datatype}{A character string representing the file type. Default value is file extension (last of all period (".") separated extensions).}

\item{seqtype}{A character string representing the sequence type. Options: "nucleotide" (default) or "amino acid".}

\item{session}{\code{session} object or character string name of \code{session} object into which NGS data should be imported. Default is "highlineR.session".}

\item{force}{A logical value specifying if previously imported file should be overwritten and re-imported.}
}
\value{
Imports specified NGS file into specified highlineR session.
}
\description{
\code{import_file} creates a Data object for an NGS file.
}
\keyword{internal}
