% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_raw_seq}
\alias{import_raw_seq}
\title{Import NGS file(s).}
\usage{
import_raw_seq(path, datatype, seqtype, session = "highlineR.session",
  force = FALSE)
}
\arguments{
\item{path}{A character string representing an absolute path to an NGS file, an absolute path to a directory of NGS files or a character vector of absolute paths to NGS files.}

\item{datatype}{A character string representing the file type. Value is inferred from file extension or specified by user as "fasta" or "fastq.}

\item{seqtype}{A character string representing the sequence type. Options: "nucleotide" (default) or "amino acid".}

\item{session}{\code{session} object or character string name of \code{session} object into which NGS data should be imported. Default is "highlineR.session".}

\item{force}{A logical value specifying if previously imported file should be overwritten and re-imported.}
}
\value{
Returns \code{session} object into which NGS data file(s) were imported.
}
\description{
\code{import_raw_seq} creates Data object(s) for NGS file(s).
}
