% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_raw_seq}
\alias{parse_raw_seq}
\alias{parse_raw_seq.session}
\alias{parse_raw_seq.fasta}
\alias{parse_raw_seq.fastq}
\alias{parse_raw_seq.csv}
\alias{parse_raw_seq.default}
\title{Parse generic}
\usage{
parse_raw_seq(x, ...)

\method{parse_raw_seq}{session}(x, encoding = "sanger", ...)

\method{parse_raw_seq}{fasta}(x, encoding = NULL, ...)

\method{parse_raw_seq}{fastq}(x, encoding = "sanger", ...)

\method{parse_raw_seq}{csv}(x, encoding = NULL, ...)

\method{parse_raw_seq}{default}(x, ...)
}
\arguments{
\item{x}{Data object or session object containing Data objects to parse.}

\item{encoding}{A character string representing the quality score encoding of FASTQ files. Options: "Sanger", "Solexa", "Illumina1.3", "Illumina1.5". "Illumina1.8".}
}
\value{
Parses Data objects by populating the object's \code{raw_seq} list with (header, sequence, [quality scores]) lists.
}
\description{
Parses the NGS file specified by the path variable of \code{Data} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{session}: Parse all Data objects in session objects.

\item \code{fasta}: Parse Data objects of FASTA NGS files.

\item \code{fastq}: Parse Data objects of FASTQ NGS files.

\item \code{csv}: Parse Data objects of custom, csv, NGS files.

\item \code{default}: default.
}}

\keyword{internal}
