% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.session}
\alias{plot.Data}
\alias{plot_init}
\alias{calc_seq_diff}
\alias{seq_simil}
\alias{sort.compressed}
\alias{calc_rel_abun}
\alias{data_melt}
\alias{aaLookup}
\alias{modetotitle}
\title{Plot generic}
\usage{
\method{plot}{session}(session, mode = "mismatch", master, sort_by,
  rf = 1, use_sample = T, ...)

\method{plot}{Data}(data, session_plot = F, mode = "mismatch",
  master = data$master, sort_by = "similarity", rf = 1,
  use_sample = T, ...)

plot_init(data, compressed, mode, master, sort_by = "similarity", rf,
  ...)

calc_seq_diff(data, compressed, mode, master, rf)

seq_simil(seq_diff)

\method{sort}{compressed}(compressed)

calc_rel_abun(compressed, seq_order)

data_melt(seq_diff, seq_order, master, ...)

aaLookup(c)

modetotitle(mode)
}
\arguments{
\item{mode}{A character string representing the desired mutation annotation for plotting. Options: "mismatch" (default), "svn" (Synonymous versus Non-Synonymous), "tvt" (Transition versus Transversion).}

\item{master}{A character string representing the sequence to which other sequences should be compared. By default, the most abundant sequence is selected.}

\item{sort_by}{A character string representing how the sequences should be ordered in a plot. Options: "similarity", "frequency".}

\item{rf}{An integer specifying which reading frame should be used when determining Synonymous vs Non-Synonymous mutations. Options: 1 (default), 2, 3.}

\item{use_sample}{A logical value specifying which environment should be plotted. If \code{True}, then the \code{sample} environment of the Data objects is plotted. If \code{False}, the complete \code{compressed} environment is plotted.}

\item{compressed}{Compressed environment of sequences for plotting.}

\item{seq_diff}{matrix of compositional differences between sequences.}

\item{seq_order}{string vector of sequences present in compressed environment ordered as desired for plotting.}

\item{c}{three-character string representing a nucleotide codon for decoding.}

\item{x}{Data object or session object containing Data objects to plot.}
}
\value{
Plots NGS data

\code{plot_init} returns a list of mutations and relative abundances of variants for plotting.

\code{calc_seq_diff} identifies compositional differences between sequences in data including annotations if applicable and stores the result in the Data object's \code{seq_diff} dataframe.

\code{seq_simil} returns a permutation (an integer vector) which rearranges the provided sequences in ascending order of number of mutations

\code{sort} returns a (an integer dataframe) which rearranges the provided sequences in ascending order of relative abundance.

\code{calc_rel_abun} returns a vector of read counts of the sequences from the \code{compressed} environment in the order specified by \code{seqs_order}.

\code{data_melt} returns a matrix of compositional differences between sequences in a format suitable for ggplot plotting.

\code{aaLookup} returns the single character code of the amino acid encoded by the codon \code{c}.

\code{modetotitle} returns string description of the mode.
}
\description{
Plots highlineR Data objects

\code{plot_init} performs plot initialization

\code{calc_seq_diff} identifies sequence differences.

\code{seq_simil} orders sequences based on similarity.

\code{sort} orders sequences based on relative abundance

\code{calc_rel_abun} calculates sequence relative abundances.

\code{data_melt} melts data for easy plotting.

\code{aaLookup} decodes triplet codon into single letter amino acid.

\code{modetotitle} converts internal mode code into verbose format for plotting.
}
\keyword{internal}
