% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{remove_Data}
\alias{remove_Data}
\title{Remove Data object from highlineR session environment.}
\usage{
remove_Data(data, session = "highlineR.session")
}
\arguments{
\item{data}{\code{Data} object or character string of absolute path of existing Data object.}

\item{session}{\code{session} object or character string name of \code{session} object containing Data object to be removed. Default is "highlineR.session".}
}
\value{
Removes specified \code{Data} object from specified \code{session}.
}
\description{
\code{remove_Data} removes a Data object from a highlineR session environment.
}
