% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{init_session}
\alias{init_session}
\title{Constructor function for S3 class representing container for Data objects.}
\usage{
init_session(session = "highlineR.session")
}
\arguments{
\item{session}{character string represting name of session to be created.}
}
\value{
Returns a constructed S3 object which inherits from the following classes:
\itemize{
\item \code{session}.
\item \code{environment}.
}
and contains an environement that will hold Data objects.
}
\description{
\code{init_session} returns constructed session object.
}
